+++ rockstarpy/__main__.py
@@ -1,0 +1,4 @@
+from rockstarpy import command_line
+
+if __name__ == '__main__':
+    command_line()
