+++ rockstarpy/command_line.py
@@ -1,12 +1,12 @@
 import argparse
-from . import rockstar
+import rockstarpy as rockstar
 
 parser = argparse.ArgumentParser(description="Python transpiler for the esoteric language Rockstar")
 parser.add_argument('input', action='store', help='Input file (.rock|.rockstar|.lyrics)')
 parser.add_argument('--output', action='store', help='Output file (.py)', default='output.py')
 parser.add_argument('-v', action='version', help='Version', version='1.2.1')
 
-def main():
+def command_line():
     args = parser.parse_args()
 
     try:
